requirements.txt
django<=1.10.3
django-filter>=0.15.3
djangorestframework>=3.5.3
requests>=2.11.1
psycopg2>=2.6.2
requests[security]
gunicorn
whitenoise